New and improved workoutLogCount

+++ workoutApp/Lift+CoreDataProperties.swift
@@ -2,7 +2,7 @@
 //  Lift+CoreDataProperties.swift
 //  
 //
-//  Created by Alexander Kvamme on 25/09/2017.
+//  Created by Alexander Kvamme on 08/11/2017.
 //
 //
 
@@ -20,7 +20,7 @@ extension Lift {
     @NSManaged public var hasBeenPerformed: Bool
     @NSManaged public var reps: Int16
     @NSManaged public var time: Int16
-    @NSManaged public var weight: Int16
+    @NSManaged public var weight: Double
     @NSManaged public var owner: ExerciseLog?
 
 }
